AMP Support code refactoring Adhere to AMP's PHPCS. Moved anonymous functions into a separate class. Move add_filter and add_action into a separate init method. Extended documentation for classes and methods. Use prepare for MySQL query. Create helper function to bootstrap Admin Support class.
